+++ lib/core/widgets/app_top_bar.dart
@@ -1,0 +1,129 @@
+import 'package:flutter/material.dart';
+import 'package:flutter_bloc/flutter_bloc.dart';
+import '../constants/app_colors.dart';
+import '../../features/auth/presentation/bloc/auth_bloc.dart';
+import '../../features/auth/presentation/bloc/auth_state.dart';
+import '../../features/notifications/notifications_page.dart';
+import '../../features/profile/profile_page.dart';
+
+class AppTopBar extends StatelessWidget {
+  final Color? backgroundColor;
+
+  const AppTopBar({super.key, this.backgroundColor});
+
+  void _navigateToProfile(BuildContext context) {
+    // Get the AuthBloc instance before navigation
+    final authBloc = context.read<AuthBloc>();
+
+    Navigator.push(
+      context,
+      MaterialPageRoute(
+        builder: (context) =>
+            BlocProvider.value(value: authBloc, child: const ProfilePage()),
+      ),
+    );
+  }
+
+  void _navigateToNotifications(BuildContext context) {
+    Navigator.push(
+      context,
+      MaterialPageRoute(builder: (context) => const NotificationsPage()),
+    );
+  }
+
+  void _onSearchTap(BuildContext context) {
+    ScaffoldMessenger.of(context).showSnackBar(
+      const SnackBar(
+        content: Text('Search functionality coming soon!'),
+        backgroundColor: AppColors.primaryColor,
+      ),
+    );
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return Container(
+      padding: const EdgeInsets.symmetric(horizontal: 16, vertical: 8),
+      decoration: BoxDecoration(color: backgroundColor ?? Colors.transparent),
+      child: Row(
+        children: [
+          // Profile Portrait
+          BlocBuilder<AuthBloc, AuthState>(
+            builder: (context, state) {
+              return GestureDetector(
+                onTap: () => _navigateToProfile(context),
+                child: Container(
+                  width: 40,
+                  height: 40,
+                  decoration: BoxDecoration(
+                    color: AppColors.primaryColor.withValues(alpha: 0.3),
+                    shape: BoxShape.circle,
+                    border: Border.all(color: Colors.white, width: 2),
+                  ),
+                  child: const Icon(
+                    Icons.person,
+                    color: Colors.white,
+                    size: 24,
+                  ),
+                ),
+              );
+            },
+          ),
+          const SizedBox(width: 16),
+
+          // Search Bar
+          Expanded(
+            child: GestureDetector(
+              onTap: () => _onSearchTap(context),
+              child: Container(
+                height: 40,
+                padding: const EdgeInsets.symmetric(horizontal: 16),
+                decoration: BoxDecoration(
+                  color: Colors.white.withValues(alpha: 0.2),
+                  borderRadius: BorderRadius.circular(20),
+                  border: Border.all(
+                    color: Colors.white.withValues(alpha: 0.3),
+                    width: 1,
+                  ),
+                ),
+                child: const Row(
+                  children: [
+                    Icon(Icons.search, color: Colors.white70, size: 20),
+                    SizedBox(width: 8),
+                    Text(
+                      'Search CritChat...',
+                      style: TextStyle(color: Colors.white70, fontSize: 16),
+                    ),
+                  ],
+                ),
+              ),
+            ),
+          ),
+          const SizedBox(width: 16),
+
+          // Notifications Button
+          GestureDetector(
+            onTap: () => _navigateToNotifications(context),
+            child: Container(
+              width: 40,
+              height: 40,
+              decoration: BoxDecoration(
+                color: Colors.white.withValues(alpha: 0.2),
+                shape: BoxShape.circle,
+                border: Border.all(
+                  color: Colors.white.withValues(alpha: 0.3),
+                  width: 1,
+                ),
+              ),
+              child: const Icon(
+                Icons.notifications_outlined,
+                color: Colors.white,
+                size: 24,
+              ),
+            ),
+          ),
+        ],
+      ),
+    );
+  }
+}
